move code to the aft package
