+++ .gitignore
@@ -2,6 +2,7 @@
 .pytest_cache/*
 __pycache__/*
 *.pyc
+venv/
 
 .idea/
